refactor: moved normalizeCollection logic to config file and created TableHead component
