+++ packages/catalyst/src/config.ts
@@ -1,9 +1,13 @@
-import { CatalystCms, UserCatalystConfig } from "./types";
+import {
+  CatalystCms,
+  CatalystConfig,
+  CatalystFields,
+  UserCatalystConfig
+} from "./types";
 import { createDashboard } from "./dashboard";
 import { createCatalystAuthObject } from "./auth";
 import { createCatalystDataObject } from "./data";
 import { createRootEndpoint } from "./endpoints";
-import { updateCollectionFields } from "./utils";
 
 export function createCatalyst<const C extends UserCatalystConfig>(
   userConfig: C
@@ -27,7 +31,7 @@ export function createCatalyst<const C extends UserCatalystConfig>(
 }
 
 function getConfigWithDefaults<const C extends UserCatalystConfig>(config: C) {
-  const updatedConfigCollections = updateCollectionFields(config);
+  const updatedConfigCollections = normalizeCollections(config);
   return {
     ...config,
     collections: updatedConfigCollections,
@@ -37,4 +41,37 @@ function getConfigWithDefaults<const C extends UserCatalystConfig>(config: C) {
       ...(config.i18n ?? {})
     }
   };
+}
+
+function normalizeCollections(config: UserCatalystConfig) {
+  const updatedCollectionFields = Object.entries(config.collections).reduce(
+    (acc, [key, collection]) => {
+      return {
+        ...acc,
+        [key]: {
+          ...collection,
+          fields: getFieldWithExposed(collection.fields)
+        }
+      };
+    },
+    {}
+  );
+  return updatedCollectionFields;
+}
+
+function getFieldWithExposed(fields: CatalystFields<CatalystConfig>) {
+  return Object.entries(fields).reduce((acc, [key, field]) => {
+    if (field.exposed === undefined) {
+      return {
+        ...acc,
+        [key]: {
+          ...field,
+          exposed: true
+        }
+      };
+    }
+    return {
+      ...acc
+    };
+  }, {});
 }
