+++ packages/catalyst/src/dashboard/routes/browse.tsx
@@ -3,6 +3,7 @@ import { CurrentSubrouteLink } from "../components/_shared/CurrentSubrouteLink";
 import { TableBody } from "../components/browse/TableBody";
 import { notFound } from "next/navigation";
 import { RouteProps } from "./types";
+import { TableHead } from "../components/browse/TableHead";
 
 export async function BrowseRoute({ cms, params, config }: RouteProps) {
   const [_, collectionName] = params;
@@ -17,26 +18,13 @@ export async function BrowseRoute({ cms, params, config }: RouteProps) {
     ([_, field]) => field.exposed
   );
 
-  function renderExposedField() {
-    if (exposedFields.length > 0) {
-      return exposedFields.map(([key, field]) => (
-        <th key={key} className="py-2 text-left font-semibold">
-          {field.label}
-        </th>
-      ));
-    }
-    return <th className="py-2 text-left font-semibold">ID</th>;
-  }
-
   return (
     <div className="flex h-full flex-col p-16">
       <h1 className="mb-8 text-4xl font-black uppercase text-red-600">
         BROWSE {collection.label}
       </h1>
       <table className="w-full table-fixed">
-        <thead>
-          <tr className="border-b border-gray-300">{renderExposedField()}</tr>
-        </thead>
+        <TableHead exposedFields={exposedFields} />
         <TableBody
           docs={docs}
           collectionName={collectionName}
